+++ plugin/src/main/java/com/skyblock/plugin/profile/ProfileRepository.java
@@ -49,7 +49,7 @@ public final class ProfileRepository {
         cfg.set("uuid", profile.getUuid().toString());
         cfg.set("purse", profile.getPurse());
         cfg.set("bank", profile.getBank());
-        for (Map.Entry<String, Long> entry : profile.getSkillXp().entrySet()) {
+        for (Map.Entry<String, Double> entry : profile.getSkillXp().entrySet()) {
             cfg.set("skills." + entry.getKey(), entry.getValue());
         }
         for (Map.Entry<String, Long> entry : profile.getCollectionXp().entrySet()) {
@@ -104,7 +104,7 @@ public final class ProfileRepository {
                     ConfigurationSection skills = cfg.getConfigurationSection("skills");
                     if (skills != null) {
                         for (String skill : skills.getKeys(false)) {
-                            profile.setSkillXp(skill, skills.getLong(skill));
+                            profile.setSkillXp(skill, skills.getDouble(skill));
                         }
                     }
